requirements.txt
streamlit
paddleocr
pillow
openpyxl
requests
numpy
